+++ module_groups/rabbitmq_cluster/outputs.tf
@@ -1,0 +1,3 @@
+output rabbitmq_autoscaling_group_name {
+  value = "${aws_cloudformation_stack.rabbitmq.outputs["AsgName"]}"
+}